+++ R/index-by.R
@@ -85,7 +85,7 @@ index_by.tbl_ts <- function(.data, ...) {
       !!! groups(.data), !! idx2, .drop = FALSE
     )
   build_tsibble(
-    tbl, key = key_data(.data), index = !! idx, index2 = !! idx2,
+    tbl, key_data = key_data(.data), index = !! idx, index2 = !! idx2,
     regular = is_regular(.data), validate = FALSE,
     ordered = is_ordered(.data), interval = interval(.data)
   )